feat(*): refactored code and output file parameter
