+++ cas2bas.py
@@ -1,0 +1,5 @@
+from cas2bas.Main import Main
+
+if __name__ == "__main__":
+    app = Main()
+    app.run()
