+++ tests/test_Cas_Format.py
@@ -1,7 +1,7 @@
 import pytest
 
-from cas2bas import dragon_tokens
-from cas2bas.cas_format import CasFormat, LEADER, SYNC, NAME_FILE_BLOCK, BASIC_FILE_IDENTIFIER, ASCII_FILE_FLAG, \
+from cas2bas import Dragon_Tokens
+from cas2bas.Cas_Format import CasFormat, LEADER, SYNC, NAME_FILE_BLOCK, BASIC_FILE_IDENTIFIER, ASCII_FILE_FLAG, \
     CONTINUOUS_FILE, DATA_BLOCK, END_OF_FILE_BLOCK, DATA_FILE_IDENTIFIER
 
 
@@ -43,7 +43,7 @@ def test_given_a_valid_byte_array_returns_a_formatted_string():
         LEADER,
         SYNC,
         END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = "10 STOP" + chr(10) + chr(13)
     header_pass = formatter.process_header()
     actual = ""
@@ -54,7 +54,7 @@ def test_given_a_valid_byte_array_returns_a_formatted_string():
 
 def test_given_a_file_without_a_sync_in_the_header_return_an_error():
     stream = [LEADER, NAME_FILE_BLOCK, 0]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = formatter.process_header()
     assert expected == actual
@@ -85,7 +85,7 @@ def test_given_a_file_without_a_data_block_sync_return_an_error():
         LEADER,
         LEADER,
         DATA_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     header_pass = formatter.process_header()
     actual = ""
@@ -131,7 +131,7 @@ def test_given_a_file_without_an_eof_block_sync_return_an_error():
         LEADER,
         LEADER,
         END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     header_pass = formatter.process_header()
     actual = ""
@@ -184,7 +184,7 @@ def test_given_a_valid_multi_block_byte_array_returns_a_formatted_string():
         LEADER,
         SYNC,
         END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = "10 STOP" + chr(10) + chr(13)
     header_pass = formatter.process_header()
     actual = ""
@@ -224,7 +224,7 @@ def test_given_a_truncated_file_returns_an_error_and_halts():
         0,
         0,
         0x0A]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     header_pass = formatter.process_header()
     if header_pass == 0:
         with pytest.raises(SystemExit) as pytest_wrapped_exception:
@@ -235,7 +235,7 @@ def test_given_a_truncated_file_returns_an_error_and_halts():
 
 def test_given_a_byte_array_without_a_name_block_returns_an_error():
     stream = [LEADER, SYNC, END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = formatter.process_header()
     assert expected == actual
@@ -256,7 +256,7 @@ def test_given_a_byte_array_without_a_basic_id_block_returns_an_error():
         0x20,
         0x20,
         DATA_FILE_IDENTIFIER]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = formatter.process_header()
     assert expected == actual
@@ -265,7 +265,7 @@ def test_given_a_byte_array_without_a_basic_id_block_returns_an_error():
 def test_given_a_byte_array_without_a_valid_format_flag_block_returns_an_error():
     stream = [LEADER, SYNC, NAME_FILE_BLOCK, 0, 0x41, 0x20, 0x20,
               0x20, 0x20, 0x20, 0x20, 0x20, BASIC_FILE_IDENTIFIER, 0x01]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = formatter.process_header()
     assert expected == actual
@@ -288,7 +288,7 @@ def test_given_a_byte_array_without_a_continuous_file_flag_block_returns_an_erro
         BASIC_FILE_IDENTIFIER,
         ASCII_FILE_FLAG,
         0x01]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = formatter.process_header()
     assert expected == actual
@@ -337,7 +337,7 @@ def test_given_a_byte_array_with_one_leader_between_data_blocks_returns_an_error
         LEADER,
         SYNC,
         END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = 0
     header_pass = formatter.process_header()
@@ -388,7 +388,7 @@ def test_given_a_byte_array_with_no_leaders_between_data_blocks_returns_an_error
         LEADER,
         SYNC,
         END_OF_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = 0
     header_pass = formatter.process_header()
@@ -435,7 +435,7 @@ def test_given_a_byte_array_without_an_end_of_file_block_returns_an_error():
         LEADER,
         SYNC,
         NAME_FILE_BLOCK]
-    formatter = CasFormat(stream, dragon_tokens.DragonToken())
+    formatter = CasFormat(stream, Dragon_Tokens.DragonToken())
     expected = -1
     actual = 0
     header_pass = formatter.process_header()
